refactor(skill-review): split into yaml/refs/types modules
Brings packages/crane-mcp/src/cli/skill-review.ts under the structural
caps (max-lines: 500, max-lines-per-function: 75, complexity: 15).

- skill-review.ts (entry, ~390 LOC): orchestrator + re-exports
- skill-review-yaml.ts: parseFrontmatter, parseYamlBlock, coerceScalar
- skill-review-refs.ts: checkReferenceValidity split into per-target helpers
- skill-review-types.ts: shared Severity, Violation, ReviewResult

Public API unchanged; test imports from './skill-review.js' work as before.
764/764 tests pass.
